// lib/types/auth.ts
export interface User {
  _id: string;
  email: string;
  full_name: string;
  role: 'student' | 'instructor' | 'admin';
}

export interface AuthState {
  user: User | null;
  isAuthenticated: boolean;
  isLoading: boolean;
}

export interface LoginCredentials {
  email: string;
  password: string;
}

export interface RegisterData {
  email: string;
  password: string;
  full_name: string;
  role: 'student' | 'instructor';
  phone: string;
}